Refactor prettier
