+++ crates/remote_server/Cargo.toml
@@ -26,6 +26,7 @@ env_logger.workspace = true
 fs.workspace = true
 futures.workspace = true
 gpui.workspace = true
+node_runtime.workspace = true
 log.workspace = true
 project.workspace = true
 remote.workspace = true
